requirements.txt
Package                       Version
----------------------------- ------------
asgiref                       3.7.2
Django                        5.0.1
django-filter                 23.5
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
pillow                        10.2.0
pip                           23.2
psycopg2-binary               2.9.9
PyJWT                         2.8.0
python-dotenv                 1.0.1
pytz                          2023.3.post1
setuptools                    68.1.2
sqlparse                      0.4.4
